chore: Add index.html for home page
This commit adds a new HTML file, index.html, to the views folder. The file contains the structure and content for the home page of the website. It includes a title, a heading, and a link to a random number page. The necessary CSS file, main.css, is also linked to the HTML file for styling purposes.