feat: Add random number page
This commit adds a new HTML file, random.html, to the views folder. The file contains the structure and content for the random number page of the website. It includes a title, a heading, a paragraph displaying the generated random number, and a button to generate a new random number. The necessary CSS file, main.css, is also linked to the HTML file for styling purposes.