refactor: Update shop route to render shop.pug with products and title
This commit modifies the shop route in the shop.routes.js file. Instead of sending the shop.html file as a response, it now renders the shop.pug file with the products and docTitle variables. This change allows for dynamic rendering of the shop page, displaying the available products and setting the page title to "Shop".