feat: Add form for adding products
This commit adds a new pug file, "add-product.pug", to the "shop/views" folder. The file contains the structure and content for the form used to add products to the website. It includes input fields for the title and a submit button. The necessary CSS files, "forms.css" and "product.css", are also linked to the pug file for styling purposes.