chore: Add head.ejs file to include HTML head tags
This commit adds a new head.ejs file to the shop/views/includes directory. The file contains the HTML markup for the head section of the website. It includes the doctype declaration, the language attribute, the charset, the viewport meta tag, the title, and a link to the main CSS file. This addition improves the code structure and allows for better organization of the HTML head content.