refactor: Add 404 and index EJS templates for better page structure
The 404.ejs and index.ejs files have been added to the shop/express-assignment-3/views directory. These templates provide the HTML markup for the 404 page and the main index page of the website. This addition improves the codebase by separating the page content into reusable templates, enhancing the maintainability and scalability of the project.