feat: Add getProducts function to shop controller
This commit adds a new function, `getProducts`, to the `shop.controller.js` file in the `shop/controllers` directory. The function is responsible for rendering the `shop` view with the appropriate data, including the `products` array, `docTitle` as 'Shop', and `path` as '/'. This addition enhances the functionality of the shop controller by providing a way to display the products on the shop page.