refactor: Update shop.routes.js to use getProducts function in shop controller
This commit updates the shop.routes.js file in the shop/routes directory to use the newly added getProducts function from the shop.controller.js file in the shop/controllers directory. The getProducts function is responsible for rendering the shop view with the appropriate data, including the products array, docTitle as 'Shop', and path as '/'. This change improves the code organization and enhances the functionality of the shop route.