refactor: Update shop.controller.js to use Product model for fetching products
This commit updates the shop.controller.js file in the shop/controllers directory to use the Product model from the product.model.js file for fetching products. Instead of importing the products array from the admin.controller.js file, the Product.fetchAll() function is now used to retrieve the products from the database. This change improves the codebase by adhering to the MVC pattern and enhances the maintainability and scalability of the project.