refactor: Use Product model for fetching products in shop.controller.js
This commit updates the shop.controller.js file in the shop/controllers directory to use the Product model for fetching products. Instead of directly accessing the products array, the Product.fetchAll() function is now used to retrieve the products from the database. This change adheres to the MVC pattern and improves the maintainability and scalability of the project.