feat: Add "Add to Cart" button to product detail page
This commit adds an "Add to Cart" button to the product detail page. The button is included within a form element and triggers a POST request to the "/cart" route when clicked. This enhancement improves the functionality of the shop by allowing users to easily add products to their cart from the product detail page.