feat: Add POST route for adding products to cart
The code changes in this commit add a POST route to the shop.routes.js file for adding products to the cart. The route is "/cart" and it calls the addToCart function in the shop.controller.js file. This enhancement improves the functionality of the shop by allowing users to easily add products to their cart.