feat: Add form for adding products to cart
This commit adds a new form in the `add-to-cart.ejs` file which allows users to add products to their cart. The form includes a "Add to Cart" button and a hidden input field to store the product ID. This enhancement improves the functionality of the shop by providing an intuitive way for users to add products to their cart.