feat: Update product-list.ejs to include add-to-cart form
The code changes in this commit update the product-list.ejs file to include an add-to-cart form for each product. The form is now included using the "add-to-cart.ejs" partial, which allows users to easily add products to their cart. This enhancement improves the functionality of the shop by providing a convenient way for users to add products to their cart directly from the product list page.