feat: Fix product not found error in getProduct function
The code changes in this commit fix a bug in the getProduct function of the shop controller. Previously, when a product was not found in the database, the function rendered a 404 page. However, this behavior has been updated to send a "Product not found!" message instead. This improvement enhances the user experience by providing a more informative error message when a product is not available.