feat: Update cart.json with increased quantity and total price
The code changes in this commit update the cart.json file to increase the quantity of the product with ID "64338234-1010-43ed-b33b-91fef69d971c" to 2 and update the total price to 79.98. This enhancement improves the functionality of the shopping cart by accurately reflecting the updated quantity and total price of the product.