chore: Add method-override middleware for HTTP method override
This commit adds the method-override middleware to the server.js file. The middleware is used to override the HTTP method based on the value of the "_method" query parameter or the "X-HTTP-Method-Override" header. This enhancement improves the functionality of the server by allowing clients to use non-standard HTTP methods for requests, such as PUT or DELETE, which are not supported by HTML forms.