refactor: Update getCart function in shop controller
This commit refactors the getCart function in the shop controller. The code changes include using the Cart.getCart method to retrieve the cart data and the Product.fetchAll method to fetch all products. The cart products are then matched with the fetched products to create an updated cart object. This improvement enhances the functionality of the cart feature in the shop.