refactor: Update getCart method in Cart class
This commit refactors the getCart method in the Cart class of cart.models.js. The changes include updating the method to use fs.readFile to read the cart data from the file and parse it into a JSON object. If there is an error reading the file, the callback is called with null. Otherwise, the callback is called with the parsed cart object. This improvement enhances the functionality of the getCart method in the Cart class.