refactor: Update shop.routes.js to include removeFromCart route
This commit updates the shop.routes.js file to include a new route for removing a product from the cart. The route is defined as a DELETE request to '/cart' and calls the removeFromCart function in the shop controller. This enhancement improves the functionality of the cart feature in the shop.