feat: Add user ID to product creation in admin.controller.js
This commit adds the user ID to the product creation process in the `addProduct` function of the `admin.controller.js` file. By including the user ID, we can associate the created product with the user who added it. This change ensures consistency with recent updates to other controllers and improves the functionality of the application.